convert (1/1): introduced prototypes for per-item filtering functions. Filtering and output on a per-entry basis rather than in a main-loop is needed as soon as we want both query + formatting.
The abook_output_item_filter is aimed to store query-compatible output filters.
For most output formats generate a strict and standard-compliant output and won't
need this.
But the goal is the have a generic customizable format compatible with an
hypothetic --query switch.

As a future example, --mutt-query could be seen as:
 --query --format="{email} {name}"

(see the 6 following patches)